SCC-2079 Add timeouts to authenticated call
It is unclear why but the method of making a HTTP request via the `Net::HTTP::start` method was intermittently timeing out. This sets all the relevant timeouts to 5 to prevent this call from hanging and causing the invocation to fail. These timeouts could hypothetically be set lower, but 5 should be a good level to ensure performance while allowing for slow responses in case of network issues etc.