SCC-2079 Make function multithreaded
This converts the main handler method for the function to a method that invokes a multi-processed pool of jobs which work on processing SHEP updates int parallel. This is a simple conversion where all the logic within each process represents the processing for a single record with the outcome being returned from the method invoked.

Hypothetically this will realize a large performance gain in the Poster as Neo4j claims to be able to handle parallel requests rather easily. But this has not been proven and we also have to make sure that the lambda is appropriately provisioned to take advantage of this.

This also adds some unit tests for the methods in the `app.rb` module