NOREF Make number of parallel processes configurable
This makes the number of parellel calls to the SHEP API configurable as an environment variable. The impetus for this is that the current setting is potentially too high and needs to be lowered. This both sets the process to 2 to do so and makes it possible to adjust the value in the AWS Lambda console without further PRs